merge

See merge request cs309/fall2021/3_rk_2!6

+++ Backend/userdb/springboot_example/src/main/java/onetoone/Reviews/Review.java
@@ -11,37 +11,34 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 
+
 @Entity
-public class Review {
-
-
+public class Review
+{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String comments;
     private LocalDateTime timeCreated = LocalDateTime.now();
 
-
     @ManyToOne
-    @JoinColumn(name = "user_id")
-
+      @JoinColumn(name = "user_id")
     private User user;
 
-    public Review(String comments) {
+    public Review(String comments) 
+    {
         this.comments = comments;
-
     }
 
-    public Review() {
+    public Review(){
     }
-
-
-
-    public int getId(){
+    
+    public Long getId()
+    {
         return id;
     }
 
-    public void setId(int id ){
+    public void setId(Long id){
         this.id = id;
     }
 
@@ -61,13 +58,13 @@ public class Review {
         this.timeCreated = timeCreated;
     }
 
-    public User getUser() {
+    public User getUser() 
+    {
         return user;
     }
 
-    public void setUser(User user) {
+    public void setUser(User user) 
+    {
         this.user = user;
     }
-
-
 }
