+++ Backend/userdb/springboot_example/src/main/java/onetoone/Roles/Role.java
@@ -14,7 +14,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String roleType;
 
     @OneToMany
@@ -22,35 +22,40 @@ public class Role {
     private List<User> users;
 
 
-
-    public Role(String roleType) {
+    public Role(String roleType) 
+    {
         this.roleType = roleType;
     }
-
-    public Role() {
+       public Role() {
     }
 
-    public int getId() {
+    public Long getId() 
+    {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) 
+    {
         this.id = id;
     }
 
-    public String getRoleType() {
+    public String getRoleType() 
+    {
         return roleType;
     }
 
-    public void setRoleType(String roleType) {
+    public void setRoleType(String roleType) 
+    {
         this.roleType = roleType;
     }
 
-    public List<User> getUsers() {
+    public List<User> getUsers() 
+    {
         return users;
     }
 
-    public void setUsers(List<User> users) {
+    public void setUsers(List<User> users) 
+    {
         this.users = users;
     }
 }
